+++ HISTORY.md
@@ -50,6 +50,7 @@
 
 ### 2022-11-07
 * objects: tweaked monitoring non-sprite morphs for development
+* blocks: op-sequence analysis, experimental
 
 ### 2022-11-02
 * updated "Codification" example project